[lib] wallet: graduate picocoin wallet handling into libccoin
Now that picocoin's stateful wallet code has been isolated from a
stateless wallet API, split off the stateless variants into libccoin.
